requirements.txt
# LangChain ecosystem
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0

# FastAPI
fastapi==0.116.1
uvicorn==0.35.0

# env
pydantic_settings==2.10.1

# HTTP
httpx==0.28.1

# SocketIO
python-socketio
python-multipart 

# MongoDB 数据库
pymongo==4.10.1

# langgraph sqlite 数据库
langgraph-checkpoint-sqlite
